Remove the `XkbStateNotify` handler
The only thing we did in response to said event is update the
`xkb_state`, but we throw that away on every keypress anyway due to the
keyeyboard layout change detection hack. This commit also removes
associated code which let us recieve and react to the event in the first
place. If a fix is ever introduced, then this commit ought to be
reverted.